+++ demo/vite.config.js
@@ -1,9 +1,4 @@
 import { defineConfig } from 'vite';
-import { terminologyVitePlugin } from '@forschungsgruppe-digital-health/terminology/vite';
-import {
-  ENABLE_PACKAGE_DISCOVERY,
-  DISCOVERY_PACKAGES
-} from './src/terminology-config.js';
 
 export default defineConfig({
   base: process.env.BASE_PATH || '/',
@@ -22,11 +17,6 @@ export default defineConfig({
   optimizeDeps: {
     force: true
   },
-  plugins: [
-    terminologyVitePlugin({
-      packages: ENABLE_PACKAGE_DISCOVERY ? DISCOVERY_PACKAGES : []
-    })
-  ],
   build: {
     outDir: 'dist',
     emptyOutDir: true,